Changed price and tax columns from float to decimal data type

+++ database/migrations/2023_09_19_011634_create_room_categories_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->integer('size');
             $table->json('type_of_beds');
             $table->integer('pax');
-            $table->float('price');
-            $table->float('tax');
+            $table->decimal('price');
+            $table->decimal('tax');
             $table->timestamps();
         });
     }
